\name{VAdata}
\alias{VAdata}

\title{VAdata: a pseudo verbal autopsy data}
\description{
  This is a pseudo sample that is constructed from a verbal autopsy
  survey and related hospital deaths. In this pseudo verbal autopsy data,
  there are 764 regiserted deaths from hospital and 738 deaths collected 
  from the general population. The total number of symptoms are 49, 
  and there are 19 causes of reported deaths. The variables names are
  region: 1 is hospital, 2 is general population.
  cod: cause of death based on hopsital death records or gold standard
  diagnostics.
  S1...S49: collected symptoms. 1, a symptom is observed, 0, not. 
}
\usage{
data(VAdata)
}

